jail: Reenable recursive use of unshare
This feature allows the caller to delay dropping permissions until it has performed some privileged
operation(s).  It can accomplish this by passing either -c or -r as the first token of the
"cmdline," followed by a script that performs the privileged operations then does: # exec unshare -U

It is implemented by bringing back the mount-moving workaround that used to allow us to call unshare
within chroot.  Unfortunately, things are not quite that simple.  First, the workaround leaves the
root directory in a bad state that cannot yet host a new user namespace; to get around this, we do a
separate unshare -R. (equivalent to chroot .) to introduce an intervening exec system call.  We
would now be done if we were willing to invoke the inner unshare without -U, but doing so would
require special logic to decide whether to pass that flag based on whether the caller wanted to do
privileged things; instead, we temporarily bind mount /proc to support custom user mappings with -U.
Now the caller can simply hijack unshare's own -c and -r (and other) flags to retain "root."